Automatic BitmapFilter creation in FilterActuator
This creates a BitmapFilter when the user tries to use `Actuate.effects(target, duration).filter(...)` without having created and added the specific filter to `target.filters` array.

It prevents a runtime error the user will get if he/she tries to use filters and tweening using the explanation provided in the readme.md. This is related to issue #36.